Avoid loading DBM-Unified as both library and check target
This is a bit ugly and isn't necessary in the normal setup because usually your library definition and your DBM-Unified workspace are the same. But in CI it always has DBM-Unified in a separate directory and then it can't know that it's the same.

Also output warnings twice because GitHub loves to hide useful
information